requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
biopython==1.79
debugpy==1.3.0
decorator==5.0.9
Django==3.2.5
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
numpy==1.21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
